requirements.txt
watchtower~=0.7.3
boto3>=1.15.0,<1.18.0
psycopg2-binary>=2.7.4
apache-airflow
apache-airflow-providers-postgres
apache-airflow-providers-amazon
